-All routes have swagger documentation -Folder refactor for module separation
