+++ src/server.ts
@@ -1,6 +1,10 @@
 import app from './index'
+import {setupSwagger} from "./swaggerConfig";
 const port = process.env.PORT || 8080;
+
+setupSwagger(app)
 
 app.listen(port, () => {
     console.log(`Running on port ${port}`);
 });
+
